+++ packages/laconia-event/src/SnsJsonInputConverter.js
@@ -1,0 +1,5 @@
+module.exports = class SnsJsonInputConverter {
+  convert(event) {
+    return JSON.parse(event.Records[0].Sns.Message);
+  }
+};