+++ packages/laconia-event/src/index.js
@@ -2,6 +2,7 @@ const S3EventInputConverter = require("./S3EventInputConverter");
 const S3StreamInputConverter = require("./S3StreamInputConverter");
 const S3JsonInputConverter = require("./S3JsonInputConverter");
 const KinesisJsonInputConverter = require("./KinesisJsonInputConverter");
+const SnsJsonInputConverter = require("./SnsJsonInputConverter");
 
 exports.s3Event = () => () => ({
   $inputConverter: new S3EventInputConverter()
@@ -17,4 +18,8 @@ exports.s3Json = () => ({ $s3 }) => ({
 
 exports.kinesisJson = () => () => ({
   $inputConverter: new KinesisJsonInputConverter()
+});
+
+exports.snsJson = () => () => ({
+  $inputConverter: new SnsJsonInputConverter()
 });